Update JavaScript definitions.
