+++ def/core-operators.ts
@@ -1,0 +1,19 @@
+export const BinaryOperators = [
+  "==", "!=", "===", "!==",
+  "<", "<=", ">", ">=",
+  "<<", ">>", ">>>",
+  "+", "-", "*", "/", "%",
+  "&",
+  "|", "^", "in",
+  "instanceof",
+];
+
+export const AssignmentOperators = [
+  "=", "+=", "-=", "*=", "/=", "%=",
+  "<<=", ">>=", ">>>=",
+  "|=", "^=", "&=",
+];
+
+export const LogicalOperators = [
+  "||", "&&",
+];
