+++ def/es2017.ts
@@ -1,0 +1,20 @@
+import { Fork } from "../types";
+import es2016Def from "./es2016";
+import typesPlugin from "../lib/types";
+import sharedPlugin from "../lib/shared";
+
+export default function (fork: Fork) {
+  fork.use(es2016Def);
+
+  const types = fork.use(typesPlugin);
+  const def = types.Type.def;
+  const defaults = fork.use(sharedPlugin).defaults;
+
+  def("Function")
+    .field("async", Boolean, defaults["false"]);
+
+  def("AwaitExpression")
+    .bases("Expression")
+    .build("argument")
+    .field("argument", def("Expression"));
+};
