+++ def/jsx.ts
@@ -1,11 +1,11 @@
 import { Fork } from "../types";
-import es7Def from "./es7";
+import es2020Def from "./es2020";
 import typesPlugin from "../lib/types";
 import sharedPlugin from "../lib/shared";
 import { namedTypes as N } from "../gen/namedTypes";
 
 export default function (fork: Fork) {
-  fork.use(es7Def);
+  fork.use(es2020Def);
 
   var types = fork.use(typesPlugin);
   var def = types.Type.def;
